requirements.txt
google-generativeai
python-dotenv
youtube-transcript-api
pypdf2
beautifulsoup4
requests